bpf: Skip ingress policy at egress of source if egress prog is in use
When a dedicated BPF prog is in use implementing ingress policy for an
endpoint, there is no need to enforce the ingress policy of the local
destination at the source. The policy will be enforced twice.

In this new mode, the identity of the source is carried in the skb mark to the
BPF program on the destination endpoint side. For this purpose, a new mark flag
is defined.

Signed-off-by: Thomas Graf <thomas@cilium.io>